pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sofa_mover"]

[project]
name = "sofa-mover"
version = "0.1.0"
description = "RL agent for the moving sofa problem"
requires-python = ">=3.10"
dependencies = [
    "torchrl",
    "jaxtyping",
    "tqdm>=4.67.3",
    "rich>=14.3.3",
    "wandb>=0.21.1",
    "skrl>=2.0.0",
]

[dependency-groups]
dev = [
    "black>=26.3.0",
    "matplotlib>=3.10.8",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
ignore = ["F722", "E402"]  # F722: jaxtyping confuses ruff, E402: warning filters before imports

[tool.mypy]
mypy_path = "src"
packages = ["sofa_mover"]
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["tensordict", "tensordict.*", "torchrl", "torchrl.*", "tqdm", "tqdm.*"]
ignore_missing_imports = true
